RELAY-195

+++ core/client/evm/eth_commit_valset.go
@@ -49,8 +49,8 @@ func (e *Client) CommitValsetHeader(
 	headerDTO := gen.ISettlementValSetHeader{
 		Version:            header.Version,
 		RequiredKeyTag:     uint8(header.RequiredKeyTag),
-		Epoch:              new(big.Int).SetUint64(header.Epoch),
-		CaptureTimestamp:   new(big.Int).SetUint64(header.CaptureTimestamp),
+		Epoch:              new(big.Int).SetUint64(uint64(header.Epoch)),
+		CaptureTimestamp:   new(big.Int).SetUint64(uint64(header.CaptureTimestamp)),
 		QuorumThreshold:    header.QuorumThreshold.Int,
 		TotalVotingPower:   header.TotalVotingPower.Int,
 		ValidatorsSszMRoot: header.ValidatorsSszMRoot,
